+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - ACMEd now displays a warning when the server indicates an error in an order or an authorization.
+- A configuration file can now include several other files.
 
 
 ## [0.4.0] - 2019-05-08